adk-orchestrator/pyproject.toml
[project]
name = "month-end-orchestrator"
version = "1.0.0"
description = "ADK-based orchestrator for Smart Month-End Close"
requires-python = ">=3.9"
dependencies = [
    "google-adk>=0.1.0",
    "google-cloud-storage>=2.14.0",
    "google-cloud-bigquery>=3.14.1",
    "google-cloud-firestore>=2.14.0",
    "google-cloud-aiplatform>=1.38.1",
    "firebase-admin>=6.2.0",
    "fastapi>=0.108.0",
    "uvicorn>=0.25.0",
    "pandas>=2.1.4",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.3",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=2.2.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["month_end_agent"]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "locust>=2.20.0",
]
